+++ src/components/ui/TabContents.tsx
@@ -11,20 +11,31 @@ export default function TabContents({ tabs, content }: TabsProps) {
 
   return (
     <div className="mt-10">
-      <div className="flex border-b border-t border-gray-300">
+      <div className="flex border-b-2 border-t-2 border-black">
         {tabs.map((tab, index) => (
-          <button
+          <div
             key={tab}
-            onClick={() => setActiveTab(index)}
-            className={`w-[50%] px-4 py-3 cursor-pointer ${
-              activeTab === index ? "text-black font-bold" : "text-black"
-            }`}
+            className="flex justify-center w-[calc(100%/3)] px-4 py-2"
           >
-            {tab}
-          </button>
+            <button
+              onClick={() => setActiveTab(index)}
+              className={`cursor-pointer px-10 py-1 rounded-lg ${
+                activeTab === index ? "bg-[#C5A76A] text-white" : "text-black"
+              }`}
+            >
+              {tab}
+            </button>
+          </div>
         ))}
       </div>
-      <div className="p-4">{content[activeTab]}</div>
+      <div className="p-4">
+        {" "}
+        {content.map((item, index) => (
+          <div key={index} hidden={index !== activeTab}>
+            {item}
+          </div>
+        ))}
+      </div>
     </div>
   );
 }
